app/requirements.txt
faiss-cpu
numpy
python-dotenv
google-generativeai
scikit-learn
jira
altair
pandas
streamlit